Stepper Motor Tutorial
This is a example for anyone interested in moving a bipolar stepper
motor for the first time.
It uses a Nema-16 Bipolar Stepper Motor, generally like $15.00 through
Amazon, Adafruit, or Sparkfun.
This example focuses on Bipolar Stepper Motors, not Unipolar Stepper
Motors. It is not necessary to use the exact motor, however users need
to be aware of voltage and current requirements of their system.
In additional requirement is an EasyDriver Stepper control board, prices
vary. It can handle up to 30V and 700mA. Turn the potentiometer on the
EasyDriver counter clockwise to reduce current, and clockwise to
increase current flow. Temperatures on the integrated circuit will rise
with increased current flow, high temperatures will cause the motor to
stall.
Power sources for this example will be a 9 Volt Batteries and a
connection through a serial cable to your computer.
If lacking the correct power source, alternatively the Arduino can
supply the power for the board and the motor itself.
Many motors can run even if they are under power as long as there is no
major load.